Add initial basic lunar simulation
