+++ lunar_rover/package.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>mars_rover</name>
+  <name>lunar_rover</name>
   <version>0.0.1</version>
-  <description>A Mars Rover Demo for SpaceROS</description>
-  <maintainer email="tonylitianyu@gmail.com">tianyu</maintainer>
+  <description>A Lunar Rover Demo for SpaceROS</description>
+  <maintainer email="jasper.grant@dal.ca">JasperGrant</maintainer>
   <license>Apache-2.0</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
